Fix ExploreSection2
